+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - Support composite command routers ([#111](https://github.com/commanded/commanded/pull/111)).
+- Aggregate state snapshots ([#121](https://github.com/commanded/commanded/pull/121)).
 
 ## v0.15.1